Move data module into separate crate
